+++ flow360/__init__.py
@@ -12,6 +12,23 @@ from .component import meshing
 from .component.case import Case
 from .component.case import CaseList as MyCases
 from .component.flow360_params import solvers
+from .component.flow360_params.boundaries import (
+    FreestreamBoundary,
+    HeatFluxWall,
+    IsothermalWall,
+    MassInflow,
+    MassOutflow,
+    NoSlipWall,
+    SlidingInterfaceBoundary,
+    SlipWall,
+    SolidAdiabaticWall,
+    SolidIsothermalWall,
+    SubsonicInflow,
+    SubsonicOutflowMach,
+    SubsonicOutflowPressure,
+    SupersonicInflow,
+    WallFunction,
+)
 from .component.flow360_params.flow360_params import (
     ActuatorDisk,
     AeroacousticOutput,
@@ -24,42 +41,28 @@ from .component.flow360_params.flow360_params import (
     Flow360MeshParams,
     Flow360Params,
     FluidDynamicsVolumeZone,
-    FreestreamBoundary,
     FreestreamFromMach,
     FreestreamFromMachReynolds,
     FreestreamFromVelocity,
     Geometry,
     HeatEquationSolver,
-    HeatFluxWall,
     HeatTransferVolumeZone,
     IncompressibleNavierStokesSolver,
     IsoSurfaceOutput,
     IsoSurfaces,
-    IsothermalWall,
     KOmegaSST,
-    MassInflow,
-    MassOutflow,
     MeshBoundary,
     MonitorOutput,
     Monitors,
     NavierStokesSolver,
     NoneSolver,
-    NoSlipWall,
     PorousMedium,
     ProbeMonitor,
     ReferenceFrame,
     SliceOutput,
     Slices,
     SlidingInterface,
-    SlidingInterfaceBoundary,
-    SlipWall,
-    SolidAdiabaticWall,
-    SolidIsothermalWall,
     SpalartAllmaras,
-    SubsonicInflow,
-    SubsonicOutflowMach,
-    SubsonicOutflowPressure,
-    SupersonicInflow,
     SurfaceIntegralMonitor,
     SurfaceOutput,
     Surfaces,
@@ -67,7 +70,6 @@ from .component.flow360_params.flow360_params import (
     UnvalidatedFlow360Params,
     VolumeOutput,
     VolumeZones,
-    WallFunction,
     ZeroFreestream,
     ZeroFreestreamFromVelocity,
     air,
