+++ flow360/component/flow360_params/boundaries.py
@@ -1,0 +1,235 @@
+"""
+Boundaries parameters
+"""
+from abc import ABCMeta
+from typing import Optional, Tuple, Union
+
+import pydantic as pd
+from pydantic import StrictStr
+
+from ..types import Axis, PositiveFloat, PositiveInt, Vector
+from .params_base import Flow360BaseModel
+from .turbulence_quantities import TurbulenceQuantitiesType
+from .unit_system import VelocityType
+
+BoundaryVelocityType = Union[VelocityType.Vector, Tuple[StrictStr, StrictStr, StrictStr]]
+BoundaryAxisType = Union[Axis, Tuple[StrictStr, StrictStr, StrictStr]]
+
+
+# pylint: enable=too-many-arguments, too-many-return-statements, too-many-branches
+class Boundary(Flow360BaseModel, metaclass=ABCMeta):
+    """Basic Boundary class"""
+
+    type: str
+    name: Optional[str] = pd.Field(
+        None, title="Name", description="Optional unique name for boundary."
+    )
+
+
+class BoundaryWithTurbulenceQuantities(Boundary, metaclass=ABCMeta):
+    """Turbulence Quantities on Boundaries"""
+
+    turbulence_quantities: Optional[TurbulenceQuantitiesType] = pd.Field(
+        alias="turbulenceQuantities"
+    )
+
+
+class NoSlipWall(Boundary):
+    """No slip wall boundary"""
+
+    type = pd.Field("NoSlipWall", const=True)
+    velocity: Optional[BoundaryVelocityType] = pd.Field(alias="Velocity")
+
+
+class SlipWall(Boundary):
+    """Slip wall boundary"""
+
+    type = pd.Field("SlipWall", const=True)
+
+
+class FreestreamBoundary(BoundaryWithTurbulenceQuantities):
+    """Freestream boundary"""
+
+    type = pd.Field("Freestream", const=True)
+    velocity: Optional[BoundaryVelocityType] = pd.Field(alias="Velocity")
+
+
+class IsothermalWall(Boundary):
+    """IsothermalWall boundary"""
+
+    type = pd.Field("IsothermalWall", const=True)
+    temperature: Union[PositiveFloat, StrictStr] = pd.Field(alias="Temperature")
+    velocity: Optional[BoundaryVelocityType] = pd.Field(alias="Velocity")
+
+
+class HeatFluxWall(Boundary):
+    """:class:`HeatFluxWall` class for specifying heat flux wall boundaries
+
+    Parameters
+    ----------
+    heatFlux : float
+        Heat flux at the wall.
+
+    velocity: BoundaryVelocityType
+        (Optional) Velocity of the wall. If not specified, the boundary is stationary.
+
+    Returns
+    -------
+    :class:`HeatFluxWall`
+        An instance of the component class HeatFluxWall.
+
+    Example
+    -------
+    >>> heatFluxWall = HeatFluxWall(heatFlux=-0.01, velocity=(0, 0, 0))
+    """
+
+    type = pd.Field("HeatFluxWall", const=True)
+    heat_flux: Union[float, StrictStr] = pd.Field(alias="heatFlux")
+    velocity: Optional[BoundaryVelocityType] = pd.Field(alias="velocity")
+
+
+class SubsonicOutflowPressure(BoundaryWithTurbulenceQuantities):
+    """SubsonicOutflowPressure boundary"""
+
+    type = pd.Field("SubsonicOutflowPressure", const=True)
+    static_pressure_ratio: PositiveFloat = pd.Field(alias="staticPressureRatio")
+
+
+class SubsonicOutflowMach(BoundaryWithTurbulenceQuantities):
+    """SubsonicOutflowMach boundary"""
+
+    type = pd.Field("SubsonicOutflowMach", const=True)
+    Mach: PositiveFloat = pd.Field(alias="MachNumber")
+
+
+class SubsonicInflow(BoundaryWithTurbulenceQuantities):
+    """SubsonicInflow boundary"""
+
+    type = pd.Field("SubsonicInflow", const=True)
+    total_pressure_ratio: PositiveFloat = pd.Field(alias="totalPressureRatio")
+    total_temperature_ratio: PositiveFloat = pd.Field(alias="totalTemperatureRatio")
+    ramp_steps: Optional[PositiveInt] = pd.Field(alias="rampSteps")
+    velocity_direction: Optional[BoundaryVelocityType] = pd.Field(alias="velocityDirection")
+
+
+class SupersonicInflow(Boundary):
+    """:class:`SupersonicInflow` class for specifying the full fluid state at supersonic inflow boundaries
+
+    Parameters
+    ----------
+    total_temperature_ratio : PositiveFloat
+        Ratio of total temperature to static temperature at the inlet.
+
+    total_pressure_ratio: PositiveFloat
+        Ratio of the total pressure to static pressure at the inlet.
+
+    static_pressure_ratio: PositiveFloat
+        Ratio of the inlet static pressure to the freestream static pressure. Default freestream static pressure in
+        Flow360 = 1.0/gamma.
+
+    velocity_direction: BoundaryAxisType
+        (Optional) 3-array of either float values or string expressions. Unit vector which specifies the direction
+        of the incoming flow. If not specified, the boundary patch normal is used to specify direction.
+
+    Returns
+    -------
+    :class:`SupersonicInflow`
+        An instance of the component class SupersonicInflow.
+
+    Example
+    -------
+    >>> supersonicInflow = SupersonifInflow(totalTemperatureRatio=2.1, totalPressureRatio=3.0, staticPressureRatio=1.2)
+    """
+
+    type = pd.Field("SupersonicInflow", const=True)
+    total_temperature_ratio: PositiveFloat = pd.Field(
+        alias="totalTemperatureRatio", supported_solver_version="release-23.3.2.0gt"
+    )
+    total_pressure_ratio: PositiveFloat = pd.Field(
+        alias="totalPressureRatio", supported_solver_version="release-23.3.2.0gt"
+    )
+    static_pressure_ratio: PositiveFloat = pd.Field(
+        alias="staticPressureRatio", supported_solver_version="release-23.3.2.0gt"
+    )
+    velocity_direction: Optional[BoundaryAxisType] = pd.Field(
+        alias="velocityDirection", supported_solver_version="release-23.3.2.0gt"
+    )
+
+
+class SlidingInterfaceBoundary(Boundary):
+    """:class: `SlidingInterface` boundary"""
+
+    type = pd.Field("SlidingInterface", const=True)
+
+
+class WallFunction(Boundary):
+    """:class: `WallFunction` boundary"""
+
+    type = pd.Field("WallFunction", const=True)
+
+
+class MassInflow(BoundaryWithTurbulenceQuantities):
+    """MassInflow boundary"""
+
+    type = pd.Field("MassInflow", const=True)
+    mass_flow_rate: PositiveFloat = pd.Field(alias="massFlowRate")
+    ramp_steps: Optional[PositiveInt] = pd.Field(alias="rampSteps")
+
+
+class MassOutflow(BoundaryWithTurbulenceQuantities):
+    """MassOutflow boundary"""
+
+    type = pd.Field("MassOutflow", const=True)
+    mass_flow_rate: PositiveFloat = pd.Field(alias="massFlowRate")
+    ramp_steps: Optional[PositiveInt] = pd.Field(alias="rampSteps")
+
+
+class SolidIsothermalWall(Boundary):
+    """:class: `SolidIsothermalWall` boundary"""
+
+    type = pd.Field("SolidIsothermalWall", const=True)
+    temperature: Union[PositiveFloat, StrictStr] = pd.Field(alias="Temperature")
+
+
+class SolidAdiabaticWall(Boundary):
+    """:class: `SolidAdiabaticWall` boundary"""
+
+    type = pd.Field("SolidAdiabaticWall", const=True)
+
+
+class TranslationallyPeriodic(Boundary):
+    """:class: `TranslationallyPeriodic` boundary"""
+
+    type = pd.Field("TranslationallyPeriodic", const=True)
+    paired_patch_name: Optional[str] = pd.Field(alias="pairedPatchName")
+    translation_vector: Optional[Vector] = pd.Field(alias="pairedPatchName")
+
+
+class RotationallyPeriodic(Boundary):
+    """:class: `RotationallyPeriodic` boundary"""
+
+    type = pd.Field("RotationallyPeriodic", const=True)
+    paired_patch_name: Optional[str] = pd.Field(alias="pairedPatchName")
+    axis_of_rotation: Optional[Vector] = pd.Field(alias="axisOfRotation")
+    theta_radians: Optional[float] = pd.Field(alias="thetaRadians")
+
+
+BoundaryType = Union[
+    NoSlipWall,
+    SlipWall,
+    FreestreamBoundary,
+    IsothermalWall,
+    HeatFluxWall,
+    SubsonicOutflowPressure,
+    SubsonicOutflowMach,
+    SubsonicInflow,
+    SupersonicInflow,
+    SlidingInterfaceBoundary,
+    WallFunction,
+    MassInflow,
+    MassOutflow,
+    SolidIsothermalWall,
+    SolidAdiabaticWall,
+    TranslationallyPeriodic,
+    RotationallyPeriodic,
+]
