+++ flow360/component/flow360_params/flow360_params.py
@@ -43,6 +43,7 @@ from ..types import (
     Vector,
 )
 from ..utils import _get_value_or_none, beta_feature
+from .boundaries import BoundaryType, WallFunction
 from .conversions import ExtraDimensionedProperty
 from .flow360_legacy import (
     LegacyModel,
@@ -95,6 +96,7 @@ from .solvers import (
     TurbulenceModelSolverTypes,
 )
 from .time_stepping import SteadyTimeStepping, TimeStepping, UnsteadyTimeStepping
+from .turbulence_quantities import TurbulenceQuantitiesType
 from .unit_system import (
     AngularVelocityType,
     AreaType,
@@ -120,9 +122,6 @@ from .unit_system import (
     unit_system_manager,
 )
 from .validations import _check_duplicate_boundary_name, _check_tri_quad_boundaries
-
-BoundaryVelocityType = Union[VelocityType.Vector, Tuple[StrictStr, StrictStr, StrictStr]]
-BoundaryAxisType = Union[Axis, Tuple[StrictStr, StrictStr, StrictStr]]
 
 
 # pylint: disable=invalid-name
@@ -151,461 +150,6 @@ class MeshBoundary(Flow360BaseModel):
     """Mesh boundary"""
 
     no_slip_walls: Union[List[str], List[int]] = pd.Field(alias="noSlipWalls")
-
-
-class TurbulentKineticEnergy(Flow360BaseModel):
-    """
-    turbulentKineticEnergy : non-dimensional [`C_inf^2`]
-        Turbulent kinetic energy. Applicable only when using SST model.
-    """
-
-    turbulent_kinetic_energy: Optional[NonNegativeFloat] = pd.Field(alias="turbulentKineticEnergy")
-
-
-class TurbulentIntensity(Flow360BaseModel):
-    """
-    turbulentIntensity : non-dimensional [`-`]
-       Turbulent intensity. Applicable only when using SST model.
-       This is related to turbulent kinetic energy as:
-       `turbulentKineticEnergy = 1.5*pow(U_ref * turbulentIntensity, 2)`.
-       Note the use of the freestream velocity U_ref instead of C_inf.
-    """
-
-    turbulent_intensity: Optional[NonNegativeFloat] = pd.Field(alias="turbulentIntensity")
-
-
-class _SpecificDissipationRate(Flow360BaseModel, metaclass=ABCMeta):
-    """
-    specificDissipationRate : non-dimensional [`C_inf/L_gridUnit`]
-        Turbulent specific dissipation rate. Applicable only when using SST model.
-    """
-
-    specific_dissipation_rate: Optional[NonNegativeFloat] = pd.Field(
-        alias="specificDissipationRate"
-    )
-
-
-class TurbulentViscosityRatio(Flow360BaseModel):
-    """
-    turbulentViscosityRatio : non-dimensional [`-`]
-        The ratio of turbulent eddy viscosity over the freestream viscosity. Applicable for both SA and SST model.
-    """
-
-    turbulent_viscosity_ratio: Optional[NonNegativeFloat] = pd.Field(
-        alias="turbulentViscosityRatio"
-    )
-
-
-class TurbulentLengthScale(Flow360BaseModel, metaclass=ABCMeta):
-    """
-    turbulentLengthScale : non-dimensional [`L_gridUnit`]
-        The turbulent length scale is an estimation of the size of the eddies that are modeled/not resolved.
-        Applicable only when using SST model. This is related to the turbulent kinetic energy and turbulent
-        specific dissipation rate as: `L_T = sqrt(k)/(pow(beta_0^*, 0.25)*w)` where `L_T` is turbulent length scale,
-        `k` is turbulent kinetic energy, `beta_0^*` is 0.09 and `w` is turbulent specific dissipation rate.
-        Applicable only when using SST model.
-    """
-
-    turbulent_length_scale: Optional[PositiveFloat] = pd.Field(alias="turbulentLengthScale")
-
-
-class ModifiedTurbulentViscosityRatio(Flow360BaseModel):
-    """
-    modifiedTurbulentViscosityRatio : non-dimensional [`-`]
-        The ratio of modified turbulent eddy viscosity (SA) over the freestream viscosity.
-        Applicable only when using SA model.
-    """
-
-    modified_turbulent_viscosity_ratio: Optional[PositiveFloat] = pd.Field(
-        alias="modifiedTurbulentViscosityRatio"
-    )
-
-
-class ModifiedTurbulentViscosity(Flow360BaseModel):
-    """
-    modifiedTurbulentViscosity : non-dimensional [`C_inf*L_gridUnit`]
-        The modified turbulent eddy viscosity (SA). Applicable only when using SA model.
-    """
-
-    modified_turbulent_viscosity: Optional[PositiveFloat] = pd.Field(
-        alias="modifiedTurbulentViscosity"
-    )
-
-
-# pylint: disable=missing-class-docstring
-class SpecificDissipationRateAndTurbulentKineticEnergy(
-    _SpecificDissipationRate, TurbulentKineticEnergy
-):
-    pass
-
-
-class TurbulentViscosityRatioAndTurbulentKineticEnergy(
-    TurbulentViscosityRatio, TurbulentKineticEnergy
-):
-    pass
-
-
-class TurbulentLengthScaleAndTurbulentKineticEnergy(TurbulentLengthScale, TurbulentKineticEnergy):
-    pass
-
-
-class TurbulentIntensityAndSpecificDissipationRate(TurbulentIntensity, _SpecificDissipationRate):
-    pass
-
-
-class TurbulentIntensityAndTurbulentViscosityRatio(TurbulentIntensity, TurbulentViscosityRatio):
-    pass
-
-
-class TurbulentIntensityAndTurbulentLengthScale(TurbulentIntensity, TurbulentLengthScale):
-    pass
-
-
-class SpecificDissipationRateAndTurbulentViscosityRatio(
-    _SpecificDissipationRate, TurbulentViscosityRatio
-):
-    pass
-
-
-class SpecificDissipationRateAndTurbulentLengthScale(
-    _SpecificDissipationRate, TurbulentLengthScale
-):
-    pass
-
-
-class TurbulentViscosityRatioAndTurbulentLengthScale(TurbulentViscosityRatio, TurbulentLengthScale):
-    pass
-
-
-# pylint: enable=missing-class-docstring
-
-TurbulenceQuantitiesType = Union[
-    TurbulentViscosityRatio,
-    TurbulentKineticEnergy,
-    TurbulentIntensity,
-    TurbulentLengthScale,
-    ModifiedTurbulentViscosityRatio,
-    ModifiedTurbulentViscosity,
-    SpecificDissipationRateAndTurbulentKineticEnergy,
-    TurbulentViscosityRatioAndTurbulentKineticEnergy,
-    TurbulentLengthScaleAndTurbulentKineticEnergy,
-    TurbulentIntensityAndSpecificDissipationRate,
-    TurbulentIntensityAndTurbulentViscosityRatio,
-    TurbulentIntensityAndTurbulentLengthScale,
-    SpecificDissipationRateAndTurbulentViscosityRatio,
-    SpecificDissipationRateAndTurbulentLengthScale,
-    TurbulentViscosityRatioAndTurbulentLengthScale,
-]
-
-
-# pylint: disable=too-many-arguments, too-many-return-statements, too-many-branches
-def TurbulenceQuantities(
-    viscosity_ratio=None,
-    modified_viscosity_ratio=None,
-    modified_viscosity=None,
-    specific_dissipation_rate=None,
-    turbulent_kinetic_energy=None,
-    turbulent_length_scale=None,
-    turbulent_intensity=None,
-):
-    """Return a matching tubulence specification object"""
-    nonNoneArgCount = sum(arg is not None for arg in locals().values())
-    if nonNoneArgCount == 0:
-        return None
-
-    if viscosity_ratio is not None:
-        if nonNoneArgCount == 1:
-            return TurbulentViscosityRatio(turbulent_viscosity_ratio=viscosity_ratio)
-        if turbulent_kinetic_energy is not None:
-            return TurbulentViscosityRatioAndTurbulentKineticEnergy(
-                turbulent_viscosity_ratio=viscosity_ratio,
-                turbulent_kinetic_energy=turbulent_kinetic_energy,
-            )
-        if turbulent_intensity is not None:
-            return TurbulentIntensityAndTurbulentViscosityRatio(
-                turbulent_viscosity_ratio=viscosity_ratio, turbulent_intensity=turbulent_intensity
-            )
-        if specific_dissipation_rate is not None:
-            return SpecificDissipationRateAndTurbulentViscosityRatio(
-                turbulent_viscosity_ratio=viscosity_ratio,
-                specific_dissipation_rate=specific_dissipation_rate,
-            )
-        if turbulent_length_scale is not None:
-            return TurbulentViscosityRatioAndTurbulentLengthScale(
-                turbulent_viscosity_ratio=viscosity_ratio,
-                turbulent_length_scale=turbulent_length_scale,
-            )
-
-    if modified_viscosity_ratio is not None and nonNoneArgCount == 1:
-        return ModifiedTurbulentViscosityRatio(
-            modified_turbulent_viscosity_ratio=modified_viscosity_ratio
-        )
-
-    if modified_viscosity is not None and nonNoneArgCount == 1:
-        return ModifiedTurbulentViscosity(modified_turbulent_viscosity=modified_viscosity)
-
-    if turbulent_intensity is not None:
-        if nonNoneArgCount == 1:
-            return TurbulentIntensity(turbulent_intensity=turbulent_intensity)
-        if specific_dissipation_rate is not None:
-            return TurbulentIntensityAndSpecificDissipationRate(
-                turbulent_intensity=turbulent_intensity,
-                specific_dissipation_rate=specific_dissipation_rate,
-            )
-        if turbulent_length_scale is not None:
-            return TurbulentIntensityAndTurbulentLengthScale(
-                turbulent_intensity=turbulent_intensity,
-                turbulent_length_scale=turbulent_length_scale,
-            )
-
-    if turbulent_kinetic_energy is not None:
-        if nonNoneArgCount == 1:
-            return TurbulentKineticEnergy(turbulent_kinetic_energy=turbulent_kinetic_energy)
-        if specific_dissipation_rate is not None:
-            return SpecificDissipationRateAndTurbulentKineticEnergy(
-                turbulent_kinetic_energy=turbulent_kinetic_energy,
-                specific_dissipation_rate=specific_dissipation_rate,
-            )
-        if turbulent_length_scale is not None:
-            return TurbulentLengthScaleAndTurbulentKineticEnergy(
-                turbulent_kinetic_energy=turbulent_kinetic_energy,
-                turbulent_length_scale=turbulent_length_scale,
-            )
-
-    if turbulent_length_scale is not None and nonNoneArgCount == 1:
-        return TurbulentLengthScale(turbulent_length_scale=turbulent_length_scale)
-
-    if specific_dissipation_rate is not None:
-        if turbulent_length_scale is not None:
-            return SpecificDissipationRateAndTurbulentLengthScale(
-                specific_dissipation_rate=specific_dissipation_rate,
-                turbulent_length_scale=turbulent_length_scale,
-            )
-
-    raise ValueError(
-        "Please recheck TurbulenceQuantities inputs and make sure they represents a valid specification."
-    )
-
-
-# pylint: enable=too-many-arguments, too-many-return-statements, too-many-branches
-
-
-class Boundary(Flow360BaseModel, metaclass=ABCMeta):
-    """Basic Boundary class"""
-
-    type: str
-    name: Optional[str] = pd.Field(
-        None, title="Name", description="Optional unique name for boundary."
-    )
-
-
-class BoundaryWithTurbulenceQuantities(Boundary, metaclass=ABCMeta):
-    """Turbulence Quantities on Boundaries"""
-
-    turbulence_quantities: Optional[TurbulenceQuantitiesType] = pd.Field(
-        alias="turbulenceQuantities"
-    )
-
-
-class NoSlipWall(Boundary):
-    """No slip wall boundary"""
-
-    type = pd.Field("NoSlipWall", const=True)
-    velocity: Optional[BoundaryVelocityType] = pd.Field(alias="Velocity")
-
-
-class SlipWall(Boundary):
-    """Slip wall boundary"""
-
-    type = pd.Field("SlipWall", const=True)
-
-
-class FreestreamBoundary(BoundaryWithTurbulenceQuantities):
-    """Freestream boundary"""
-
-    type = pd.Field("Freestream", const=True)
-    velocity: Optional[BoundaryVelocityType] = pd.Field(alias="Velocity")
-
-
-class IsothermalWall(Boundary):
-    """IsothermalWall boundary"""
-
-    type = pd.Field("IsothermalWall", const=True)
-    temperature: Union[PositiveFloat, StrictStr] = pd.Field(alias="Temperature")
-    velocity: Optional[BoundaryVelocityType] = pd.Field(alias="Velocity")
-
-
-class HeatFluxWall(Boundary):
-    """:class:`HeatFluxWall` class for specifying heat flux wall boundaries
-
-    Parameters
-    ----------
-    heatFlux : float
-        Heat flux at the wall.
-
-    velocity: BoundaryVelocityType
-        (Optional) Velocity of the wall. If not specified, the boundary is stationary.
-
-    Returns
-    -------
-    :class:`HeatFluxWall`
-        An instance of the component class HeatFluxWall.
-
-    Example
-    -------
-    >>> heatFluxWall = HeatFluxWall(heatFlux=-0.01, velocity=(0, 0, 0))
-    """
-
-    type = pd.Field("HeatFluxWall", const=True)
-    heat_flux: Union[float, StrictStr] = pd.Field(alias="heatFlux")
-    velocity: Optional[BoundaryVelocityType] = pd.Field(alias="velocity")
-
-
-class SubsonicOutflowPressure(BoundaryWithTurbulenceQuantities):
-    """SubsonicOutflowPressure boundary"""
-
-    type = pd.Field("SubsonicOutflowPressure", const=True)
-    static_pressure_ratio: PositiveFloat = pd.Field(alias="staticPressureRatio")
-
-
-class SubsonicOutflowMach(BoundaryWithTurbulenceQuantities):
-    """SubsonicOutflowMach boundary"""
-
-    type = pd.Field("SubsonicOutflowMach", const=True)
-    Mach: PositiveFloat = pd.Field(alias="MachNumber")
-
-
-class SubsonicInflow(BoundaryWithTurbulenceQuantities):
-    """SubsonicInflow boundary"""
-
-    type = pd.Field("SubsonicInflow", const=True)
-    total_pressure_ratio: PositiveFloat = pd.Field(alias="totalPressureRatio")
-    total_temperature_ratio: PositiveFloat = pd.Field(alias="totalTemperatureRatio")
-    ramp_steps: Optional[PositiveInt] = pd.Field(alias="rampSteps")
-    velocity_direction: Optional[BoundaryVelocityType] = pd.Field(alias="velocityDirection")
-
-
-class SupersonicInflow(Boundary):
-    """:class:`SupersonicInflow` class for specifying the full fluid state at supersonic inflow boundaries
-
-    Parameters
-    ----------
-    total_temperature_ratio : PositiveFloat
-        Ratio of total temperature to static temperature at the inlet.
-
-    total_pressure_ratio: PositiveFloat
-        Ratio of the total pressure to static pressure at the inlet.
-
-    static_pressure_ratio: PositiveFloat
-        Ratio of the inlet static pressure to the freestream static pressure. Default freestream static pressure in
-        Flow360 = 1.0/gamma.
-
-    velocity_direction: BoundaryAxisType
-        (Optional) 3-array of either float values or string expressions. Unit vector which specifies the direction
-        of the incoming flow. If not specified, the boundary patch normal is used to specify direction.
-
-    Returns
-    -------
-    :class:`SupersonicInflow`
-        An instance of the component class SupersonicInflow.
-
-    Example
-    -------
-    >>> supersonicInflow = SupersonifInflow(totalTemperatureRatio=2.1, totalPressureRatio=3.0, staticPressureRatio=1.2)
-    """
-
-    type = pd.Field("SupersonicInflow", const=True)
-    total_temperature_ratio: PositiveFloat = pd.Field(
-        alias="totalTemperatureRatio", supported_solver_version="release-23.3.2.0gt"
-    )
-    total_pressure_ratio: PositiveFloat = pd.Field(
-        alias="totalPressureRatio", supported_solver_version="release-23.3.2.0gt"
-    )
-    static_pressure_ratio: PositiveFloat = pd.Field(
-        alias="staticPressureRatio", supported_solver_version="release-23.3.2.0gt"
-    )
-    velocity_direction: Optional[BoundaryAxisType] = pd.Field(
-        alias="velocityDirection", supported_solver_version="release-23.3.2.0gt"
-    )
-
-
-class SlidingInterfaceBoundary(Boundary):
-    """:class: `SlidingInterface` boundary"""
-
-    type = pd.Field("SlidingInterface", const=True)
-
-
-class WallFunction(Boundary):
-    """:class: `WallFunction` boundary"""
-
-    type = pd.Field("WallFunction", const=True)
-
-
-class MassInflow(BoundaryWithTurbulenceQuantities):
-    """MassInflow boundary"""
-
-    type = pd.Field("MassInflow", const=True)
-    mass_flow_rate: PositiveFloat = pd.Field(alias="massFlowRate")
-    ramp_steps: Optional[PositiveInt] = pd.Field(alias="rampSteps")
-
-
-class MassOutflow(BoundaryWithTurbulenceQuantities):
-    """MassOutflow boundary"""
-
-    type = pd.Field("MassOutflow", const=True)
-    mass_flow_rate: PositiveFloat = pd.Field(alias="massFlowRate")
-    ramp_steps: Optional[PositiveInt] = pd.Field(alias="rampSteps")
-
-
-class SolidIsothermalWall(Boundary):
-    """:class: `SolidIsothermalWall` boundary"""
-
-    type = pd.Field("SolidIsothermalWall", const=True)
-    temperature: Union[PositiveFloat, StrictStr] = pd.Field(alias="Temperature")
-
-
-class SolidAdiabaticWall(Boundary):
-    """:class: `SolidAdiabaticWall` boundary"""
-
-    type = pd.Field("SolidAdiabaticWall", const=True)
-
-
-class TranslationallyPeriodic(Boundary):
-    """:class: `TranslationallyPeriodic` boundary"""
-
-    type = pd.Field("TranslationallyPeriodic", const=True)
-    paired_patch_name: Optional[str] = pd.Field(alias="pairedPatchName")
-    translation_vector: Optional[Vector] = pd.Field(alias="pairedPatchName")
-
-
-class RotationallyPeriodic(Boundary):
-    """:class: `RotationallyPeriodic` boundary"""
-
-    type = pd.Field("RotationallyPeriodic", const=True)
-    paired_patch_name: Optional[str] = pd.Field(alias="pairedPatchName")
-    axis_of_rotation: Optional[Vector] = pd.Field(alias="axisOfRotation")
-    theta_radians: Optional[float] = pd.Field(alias="thetaRadians")
-
-
-BoundaryType = Union[
-    NoSlipWall,
-    SlipWall,
-    FreestreamBoundary,
-    IsothermalWall,
-    HeatFluxWall,
-    SubsonicOutflowPressure,
-    SubsonicOutflowMach,
-    SubsonicInflow,
-    SupersonicInflow,
-    SlidingInterfaceBoundary,
-    WallFunction,
-    MassInflow,
-    MassOutflow,
-    SolidIsothermalWall,
-    SolidAdiabaticWall,
-    TranslationallyPeriodic,
-    RotationallyPeriodic,
-]
 
 
 class ForcePerArea(Flow360BaseModel):
