+++ tests/params/test_params_boundary.py
@@ -4,25 +4,30 @@ import pydantic as pd
 import pytest
 
 import flow360 as fl
-from flow360.component.flow360_params.flow360_params import (
-    Flow360Params,
+from flow360.component.flow360_params.boundaries import (
     FreestreamBoundary,
     HeatFluxWall,
     IsothermalWall,
     MassInflow,
     MassOutflow,
-    MeshBoundary,
     NoSlipWall,
     SlidingInterfaceBoundary,
     SlipWall,
     SolidAdiabaticWall,
     SolidIsothermalWall,
-    SteadyTimeStepping,
     SubsonicInflow,
     SubsonicOutflowMach,
     SubsonicOutflowPressure,
-    TurbulenceQuantities,
+    SupersonicInflow,
     WallFunction,
+)
+from flow360.component.flow360_params.flow360_params import (
+    Flow360Params,
+    MeshBoundary,
+    SteadyTimeStepping,
+)
+from flow360.component.flow360_params.turbulence_quantities import (
+    get_turbulence_quantities,
 )
 from flow360.exceptions import Flow360ValidationError
 from tests.utils import compare_to_ref, to_file_from_file_test
@@ -220,7 +225,9 @@ def test_boundary_types():
     bc = SubsonicOutflowMach(
         name="SomeBC",
         Mach=0.2,
-        turbulence_quantities=TurbulenceQuantities(turbulent_intensity=0.2, viscosity_ratio=10),
+        turbulence_quantities=get_turbulence_quantities(
+            turbulent_intensity=0.2, viscosity_ratio=10
+        ),
     )
 
     assert bc.turbulence_quantities.turbulent_intensity == 0.2
@@ -229,14 +236,14 @@ def test_boundary_types():
     bc = SubsonicOutflowPressure(
         name="SomeBC",
         static_pressure_ratio=0.2,
-        turbulence_quantities=TurbulenceQuantities(),
+        turbulence_quantities=get_turbulence_quantities(),
     )
 
     assert bc.turbulence_quantities is None
 
     bc = FreestreamBoundary(
         name="SomeBC",
-        turbulence_quantities=TurbulenceQuantities(viscosity_ratio=14),
+        turbulence_quantities=get_turbulence_quantities(viscosity_ratio=14),
     )
 
     assert bc.turbulence_quantities.turbulent_viscosity_ratio == 14
@@ -244,7 +251,7 @@ def test_boundary_types():
     bc = SubsonicOutflowPressure(
         name="SomeBC",
         static_pressure_ratio=0.2,
-        turbulence_quantities=TurbulenceQuantities(
+        turbulence_quantities=get_turbulence_quantities(
             viscosity_ratio=124, turbulent_kinetic_energy=0.2
         ),
     )
@@ -255,7 +262,7 @@ def test_boundary_types():
     bc = SubsonicOutflowMach(
         name="SomeBC",
         Mach=0.2,
-        turbulence_quantities=TurbulenceQuantities(
+        turbulence_quantities=get_turbulence_quantities(
             specific_dissipation_rate=124, viscosity_ratio=0.2
         ),
     )
@@ -267,7 +274,9 @@ def test_boundary_types():
         name="SomeBC",
         total_pressure_ratio=0.2,
         total_temperature_ratio=0.43,
-        turbulence_quantities=TurbulenceQuantities(viscosity_ratio=124, turbulent_length_scale=1.2),
+        turbulence_quantities=get_turbulence_quantities(
+            viscosity_ratio=124, turbulent_length_scale=1.2
+        ),
     )
 
     assert bc.turbulence_quantities.turbulent_viscosity_ratio == 124
@@ -276,7 +285,7 @@ def test_boundary_types():
     bc = MassInflow(
         name="SomeBC",
         mass_flow_rate=0.2,
-        turbulence_quantities=TurbulenceQuantities(modified_viscosity_ratio=1.2),
+        turbulence_quantities=get_turbulence_quantities(modified_viscosity_ratio=1.2),
     )
 
     assert bc.turbulence_quantities.modified_turbulent_viscosity_ratio == 1.2
@@ -284,7 +293,7 @@ def test_boundary_types():
     bc = MassInflow(
         name="SomeBC",
         mass_flow_rate=0.2,
-        turbulence_quantities=TurbulenceQuantities(turbulent_intensity=0.2),
+        turbulence_quantities=get_turbulence_quantities(turbulent_intensity=0.2),
     )
 
     assert bc.turbulence_quantities.turbulent_intensity == 0.2
@@ -292,7 +301,7 @@ def test_boundary_types():
     bc = MassInflow(
         name="SomeBC",
         mass_flow_rate=0.2,
-        turbulence_quantities=TurbulenceQuantities(turbulent_kinetic_energy=12.2),
+        turbulence_quantities=get_turbulence_quantities(turbulent_kinetic_energy=12.2),
     )
 
     assert bc.turbulence_quantities.turbulent_kinetic_energy == 12.2
@@ -300,7 +309,7 @@ def test_boundary_types():
     bc = MassInflow(
         name="SomeBC",
         mass_flow_rate=0.2,
-        turbulence_quantities=TurbulenceQuantities(turbulent_length_scale=1.23),
+        turbulence_quantities=get_turbulence_quantities(turbulent_length_scale=1.23),
     )
 
     assert bc.turbulence_quantities.turbulent_length_scale == 1.23
@@ -308,7 +317,7 @@ def test_boundary_types():
     bc = MassOutflow(
         name="SomeBC",
         mass_flow_rate=0.2,
-        turbulence_quantities=TurbulenceQuantities(modified_viscosity=1.2),
+        turbulence_quantities=get_turbulence_quantities(modified_viscosity=1.2),
     )
 
     assert bc.turbulence_quantities.modified_turbulent_viscosity == 1.2
@@ -316,7 +325,7 @@ def test_boundary_types():
     bc = MassOutflow(
         name="SomeBC",
         mass_flow_rate=0.2,
-        turbulence_quantities=TurbulenceQuantities(
+        turbulence_quantities=get_turbulence_quantities(
             turbulent_intensity=0.88, specific_dissipation_rate=100
         ),
     )
@@ -327,7 +336,7 @@ def test_boundary_types():
     bc = MassOutflow(
         name="SomeBC",
         mass_flow_rate=0.2,
-        turbulence_quantities=TurbulenceQuantities(
+        turbulence_quantities=get_turbulence_quantities(
             turbulent_intensity=0.88, turbulent_length_scale=10
         ),
     )
@@ -338,7 +347,7 @@ def test_boundary_types():
     bc = MassOutflow(
         name="SomeBC",
         mass_flow_rate=0.2,
-        turbulence_quantities=TurbulenceQuantities(
+        turbulence_quantities=get_turbulence_quantities(
             turbulent_kinetic_energy=0.88, specific_dissipation_rate=10
         ),
     )
@@ -349,7 +358,7 @@ def test_boundary_types():
     bc = MassOutflow(
         name="SomeBC",
         mass_flow_rate=0.2,
-        turbulence_quantities=TurbulenceQuantities(
+        turbulence_quantities=get_turbulence_quantities(
             turbulent_kinetic_energy=0.88, specific_dissipation_rate=10
         ),
     )
@@ -360,7 +369,7 @@ def test_boundary_types():
     bc = MassOutflow(
         name="SomeBC",
         mass_flow_rate=0.2,
-        turbulence_quantities=TurbulenceQuantities(
+        turbulence_quantities=get_turbulence_quantities(
             turbulent_kinetic_energy=0.88, turbulent_length_scale=10
         ),
     )
@@ -371,7 +380,7 @@ def test_boundary_types():
     bc = MassOutflow(
         name="SomeBC",
         mass_flow_rate=0.2,
-        turbulence_quantities=TurbulenceQuantities(
+        turbulence_quantities=get_turbulence_quantities(
             specific_dissipation_rate=0.88, turbulent_length_scale=10
         ),
     )
@@ -383,7 +392,7 @@ def test_boundary_types():
         MassOutflow(
             name="SomeBC",
             mass_flow_rate=0.2,
-            turbulence_quantities=TurbulenceQuantities(
+            turbulence_quantities=get_turbulence_quantities(
                 specific_dissipation_rate=0.88, modified_viscosity=10
             ),
         )
@@ -392,7 +401,7 @@ def test_boundary_types():
         MassOutflow(
             name="SomeBC",
             mass_flow_rate=0.2,
-            turbulence_quantities=TurbulenceQuantities(specific_dissipation_rate=0.88),
+            turbulence_quantities=get_turbulence_quantities(specific_dissipation_rate=0.88),
         )
