+++ tests/test_volume_mesh.py
@@ -1,11 +1,11 @@
 import pytest
 
 import flow360
+from flow360.component.flow360_params.boundaries import NoSlipWall
 from flow360.component.flow360_params.flow360_params import (
     Flow360MeshParams,
     Flow360Params,
     MeshBoundary,
-    NoSlipWall,
 )
 from flow360.component.volume_mesh import (
     CompressionFormat,
